Refactoring: change structure
